Removed PlayOutRequiredResponsesAction class
